Add files via upload
In order to promote the performance of infrared and visual image fusion and provide better visual effects, this paper proposes a hybrid fusion method for infrared and visual image by the combination of discrete stationary wavelet transform (DSWT), discrete cosine transform (DCT) and local spatial frequency (LSF).
The proposed method has three key processing steps. Firstly, DSWT is employed to decompose the
important features of the source image into a series of sub-images with different levels and spatial frequencies. Secondly, DCT is used to separate the significant details of the sub-images according to the
energy of different frequencies. Thirdly, LSF is applied to enhance the regional features of DCT coefficients, and it can be helpful and useful for image feature extraction. Some frequently-used image fusion
methods and evaluation metrics are employed to evaluate the validity of the proposed method. The
experiments indicate that the proposed method can achieve good fusion effect, and it is more efficient
than other conventional image fusion methods.